Refactor, separate functions into files
